예약 수정 및 취소 시 요청한 회원에 대한 검증 로직 추가

+++ src/main/java/com/sweetypie/sweetypie/service/ReservationService.java
@@ -55,11 +55,15 @@ public class ReservationService {
         return checkDuplicateReservationDate(bookedDates, reservationDto, member, accommodation);
     }
 
-    public Reservation updateReservation(Long reservationId, ReservationDto reservationDto) {
+    public Reservation updateReservation(Long reservationId, Long memberId, ReservationDto reservationDto) {
 
         handleCheckoutBeforeCheckInInputException(reservationDto);
 
         Reservation reservation = reservationRepository.findById(reservationId).orElseThrow(() -> new DataNotFoundException("예약 내역을 찾을 수 없습니다."));
+
+        if (!memberId.equals(reservation.getMember().getId())) {
+            throw new InputNotValidException("요청한 회원정보와 예약된 회원정보가 일치하지 않습니다.");
+        }
 
         validateTotalPrice(reservationDto, reservation.getAccommodation().getPrice());
 
@@ -71,7 +75,13 @@ public class ReservationService {
 
     }
 
-    public void deleteReservation(Long reservationId) {
+    public void deleteReservation(Long reservationId, Long memberId) {
+
+        Reservation reservation = reservationRepository.findById(reservationId).orElseThrow(() -> new DataNotFoundException("예약 내역을 찾을 수 없습니다."));
+
+        if (!memberId.equals(reservation.getMember().getId())) {
+            throw new InputNotValidException("요청한 회원정보와 예약된 회원정보가 일치하지 않습니다.");
+        }
 
         reservationRepository.deleteById(reservationId);
     }
